Stworzone ścieżki do ankiet
