+++ src/app/mock-survey.ts
@@ -1,7 +1,7 @@
 import {Survey} from './models/survey';
 
 export const SURVEYS: Survey[] = [
-  {name: 'Semestr letni 2020', description: 'Oceń organizacje semestru 2020'},
-  {name: 'Juwenalia 2020', description: 'Oceń organizacje Juwenaliów 2020'},
-  {name: 'Oceń naszą aplikację', description: 'Oceń co podoba wam się w naszej aplikacji!'}
+  {id: 1, name: 'Semestr letni 2020', description: 'Oceń organizacje semestru 2020'},
+  {id: 2, name: 'Juwenalia 2020', description: 'Oceń organizacje Juwenaliów 2020'},
+  {id: 3, name: 'Oceń naszą aplikację', description: 'Oceń co podoba wam się w naszej aplikacji!'}
 ];
