+++ src/app/models/survey.ts
@@ -1,4 +1,5 @@
 export class Survey {
+  id: number;
   name: string;
   description: string;
