--- src/app/surveys/add-edit.component.html
@@ -1,51 +1,0 @@
-﻿<h1 *ngIf="isAddMode">Add User</h1>
-<h1 *ngIf="!isAddMode">Edit User</h1>
-<form [formGroup]="form" (ngSubmit)="onSubmit()">
-  <div class="form-row">
-    <div class="form-group col">
-      <label for="firstName">Imię</label>
-      <input type="text" formControlName="firstName" class="form-control"
-             [ngClass]="{ 'is-invalid': submitted && f.firstName.errors }"/>
-      <div *ngIf="submitted && f.firstName.errors" class="invalid-feedback">
-        <div *ngIf="f.firstName.errors.required">Imię jest wymagane</div>
-      </div>
-    </div>
-    <div class="form-group col">
-      <label for="lastName">Nazwisko</label>
-      <input type="text" formControlName="lastName" class="form-control"
-             [ngClass]="{ 'is-invalid': submitted && f.lastName.errors }"/>
-      <div *ngIf="submitted && f.lastName.errors" class="invalid-feedback">
-        <div *ngIf="f.lastName.errors.required">Nazwisko jest wymagane</div>
-      </div>
-    </div>
-  </div>
-  <div class="form-row">
-    <div class="form-group col">
-      <label for="username">Nazwa użytkownika</label>
-      <input type="text" formControlName="username" class="form-control"
-             [ngClass]="{ 'is-invalid': submitted && f.username.errors }"/>
-      <div *ngIf="submitted && f.username.errors" class="invalid-feedback">
-        <div *ngIf="f.username.errors.required">Nazwa użytkownika jest wymagana</div>
-      </div>
-    </div>
-    <div class="form-group col">
-      <label for="password">
-        Hasło
-        <em *ngIf="!isAddMode">(Zostaw puste aby zachować takie samo hasło)</em>
-      </label>
-      <input type="password" formControlName="password" class="form-control"
-             [ngClass]="{ 'is-invalid': submitted && f.password.errors }"/>
-      <div *ngIf="submitted && f.password.errors" class="invalid-feedback">
-        <div *ngIf="f.password.errors.required">Hasło jest wymagane</div>
-        <div *ngIf="f.password.errors.minlength">Hasło powinno mieć przynajmniej 6 liter</div>
-      </div>
-    </div>
-  </div>
-  <div class="form-group">
-    <button [disabled]="loading" class="btn btn-primary">
-      <span *ngIf="loading" class="spinner-border spinner-border-sm mr-1"></span>
-      Zapisz
-    </button>
-    <a routerLink="/users" class="btn btn-link">Anuluj</a>
-  </div>
-</form>
