+++ src/app/surveys/create.component.html
@@ -1,0 +1,1 @@
+﻿<h1>Stwórz Ankietę</h1>
