+++ src/app/surveys/layout.component.html
@@ -1,5 +1,5 @@
 ﻿<div class="p-4">
-    <div class="container">
-        <router-outlet></router-outlet>
-    </div>
+  <div class="container">
+    <router-outlet></router-outlet>
+  </div>
 </div>
