+++ src/app/surveys/layout.component.ts
@@ -1,4 +1,5 @@
-﻿import { Component } from '@angular/core';
+﻿import {Component} from '@angular/core';
 
-@Component({ templateUrl: 'layout.component.html' })
-export class LayoutComponent { }
+@Component({templateUrl: 'layout.component.html'})
+export class LayoutComponent {
+}
