+++ src/app/surveys/surveys-routing.module.ts
@@ -3,15 +3,16 @@ import {Routes, RouterModule} from '@angular/router';
 
 import {LayoutComponent} from './layout.component';
 import {ListComponent} from './list.component';
-import {AddEditComponent} from './add-edit.component';
+import {CreateComponent} from './create.component';
+import {FillComponent} from './fill.component';
 
 const routes: Routes = [
   {
     path: '', component: LayoutComponent,
     children: [
       {path: '', component: ListComponent},
-      {path: 'add', component: AddEditComponent},
-      {path: 'edit/:id', component: AddEditComponent}
+      {path: 'create', component: CreateComponent},
+      {path: 'fill/:id', component: FillComponent}
     ]
   }
 ];
